+++ craft/src/test/java/io/xj/craft/foundation/CraftFoundationInitialIT.java
@@ -136,11 +136,11 @@ public class CraftFoundationInitialIT {
     Iterator<LinkChord> it = resultLinkChords.iterator();
 
     LinkChord chordOne = it.next();
-    assertEquals(Double.valueOf(0), chordOne.getPosition());
+    assertEquals(Integer.valueOf(0), chordOne.getPosition());
     assertEquals("C minor", chordOne.getName());
 
     LinkChord chordTwo = it.next();
-    assertEquals(Double.valueOf(8), chordTwo.getPosition());
+    assertEquals(Integer.valueOf(8), chordTwo.getPosition());
     assertEquals("Db minor", chordTwo.getName());
