+++ craft/src/test/java/io/xj/craft/foundation/CraftFoundationNextMainIT.java
@@ -3,8 +3,6 @@ package io.xj.craft.foundation;
 
 import io.xj.core.CoreModule;
 import io.xj.core.access.impl.Access;
-import io.xj.core.model.phase.PhaseType;
-import io.xj.craft.CraftFactory;
 import io.xj.core.dao.ChoiceDAO;
 import io.xj.core.dao.LinkChordDAO;
 import io.xj.core.dao.LinkDAO;
@@ -18,13 +16,13 @@ import io.xj.core.model.link.LinkState;
 import io.xj.core.model.link_chord.LinkChord;
 import io.xj.core.model.link_meme.LinkMeme;
 import io.xj.core.model.pattern.PatternType;
+import io.xj.core.model.phase.PhaseType;
 import io.xj.core.model.user_role.UserRoleType;
 import io.xj.core.testing.Testing;
 import io.xj.core.work.basis.Basis;
 import io.xj.core.work.basis.BasisFactory;
+import io.xj.craft.CraftFactory;
 import io.xj.craft.CraftModule;
-
-
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
@@ -177,11 +175,11 @@ public class CraftFoundationNextMainIT {
     Iterator<LinkChord> it = resultLinkChords.iterator();
 
     LinkChord chordOne = it.next();
-    assertEquals(Double.valueOf(0), chordOne.getPosition());
+    assertEquals(Integer.valueOf(0), chordOne.getPosition());
     assertEquals("G minor", chordOne.getName());
 
     LinkChord chordTwo = it.next();
-    assertEquals(Double.valueOf(8), chordTwo.getPosition());
+    assertEquals(Integer.valueOf(8), chordTwo.getPosition());
     assertEquals("Ab minor", chordTwo.getName());
 
     // choice of macro-type pattern
